+++ C/socket_utils.c
@@ -26,7 +26,7 @@ struct sockaddr_in server;
 int demo_socket_;
 struct sockaddr_in demo_server;
 
-// static int sequential_send(double* data, int size);
+// static int sequential_send(float* data, int size);
 static matrix_t* sequential_read(int size);
 
 int init_connection(int port) {
@@ -66,18 +66,18 @@ int init_demo_connection(int port) {
 
 experience_buffer* build_demo_buffer(int size, int transition_dim) {
   experience_buffer* ret = init_experience_buffer(size);
-  double signals[DEMO_INFO_DIM];
-  double reply[transition_dim];
+  float signals[DEMO_INFO_DIM];
+  float reply[transition_dim];
   signals[0] = 1;
   for (int i = 0; i < size; ++i) {
-    int w = write(demo_socket_, signals, DEMO_INFO_DIM*sizeof(double));
-    if (w != DEMO_INFO_DIM*sizeof(double)) {
+    int w = write(demo_socket_, signals, DEMO_INFO_DIM*sizeof(float));
+    if (w != DEMO_INFO_DIM*sizeof(float)) {
       printf("[SEND] Failed to send full packet\n");
       exit(1);
     }
     
-    int r = read(demo_socket_, reply, transition_dim*sizeof(double));
-    if (r != transition_dim*sizeof(double)) {
+    int r = read(demo_socket_, reply, transition_dim*sizeof(float));
+    if (r != transition_dim*sizeof(float)) {
       printf("[SEND] Failed to read full response\n");
       exit(1);
     }
@@ -101,24 +101,24 @@ matrix_t* sim_send(matrix_t* t, int* flag, int state_dim, int act_dim) {
   
   int send_size = act_dim  + FLAG_DIM;
   int receive_size = state_dim + INFO_DIM;
-  // double server_reply[receive_size];
-  double packet[send_size];
-  packet[0] = (double)flag[0];
-  packet[1] = (double)flag[1];
+  // float server_reply[receive_size];
+  float packet[send_size];
+  packet[0] = (float)flag[0];
+  packet[1] = (float)flag[1];
   free(flag);
   for (int i = 2; i < send_size; ++i) {
     packet[i] = t->data[i-2];
   }
   // printf("sending %d\n", send_size);
-  int w = write(socket_, packet, send_size*sizeof(double));
-  if (w != send_size*sizeof(double)) {
+  int w = write(socket_, packet, send_size*sizeof(float));
+  if (w != send_size*sizeof(float)) {
     printf("[SEND] Failed to send full packet\n");
     exit(1);
   }
   //printf("sent %d", send_size);
-  // int r = read(socket_, server_reply, receive_size*sizeof(double));
-  // if (r != receive_size*sizeof(double)) {
-  //   printf("[SEND] Failed to read full response expecting %d got %d\n", receive_size, r/sizeof(double));
+  // int r = read(socket_, server_reply, receive_size*sizeof(float));
+  // if (r != receive_size*sizeof(float)) {
+  //   printf("[SEND] Failed to read full response expecting %d got %d\n", receive_size, r/sizeof(float));
   //   exit(1);
   // }
   // matrix_t* ret = new_matrix(1, receive_size);
@@ -131,13 +131,13 @@ void close_connection() {
   close(socket_);
 }
 
-// static int sequential_send(double* data, int size);
+// static int sequential_send(float* data, int size);
 // static matrix_t* sequentail_read(int size);
 
-// static int sequential_send(double* data, int size) {
+// static int sequential_send(float* data, int size) {
 //   int sent = 0;
 
-//   double packet[BLOCK_SIZE];
+//   float packet[BLOCK_SIZE];
 
 //   for ()
 
@@ -148,15 +148,15 @@ void close_connection() {
 static matrix_t* sequential_read(int size) {
   int read_num = 0;
   matrix_t* ret = new_matrix(1, size);
-  double server_reply[BLOCK_SIZE];
+  float server_reply[BLOCK_SIZE];
   while (read_num < size) {
     int nxt_block = size - read_num > BLOCK_SIZE ? BLOCK_SIZE : size - read_num;
-    int r = read(socket_, server_reply, nxt_block*sizeof(double));
-    if (r != nxt_block*sizeof(double)) {
+    int r = read(socket_, server_reply, nxt_block*sizeof(float));
+    if (r != nxt_block*sizeof(float)) {
       printf("[SEND] Failed to read full response expecting %d got %d\n", size, r);
       exit(1);
     }
-    memcpy(ret->data+read_num, server_reply, nxt_block*sizeof(double));
+    memcpy(ret->data+read_num, server_reply, nxt_block*sizeof(float));
     read_num += nxt_block;
   }
   return ret;
